Added IsValidEntityCheck to SoundHook. Added optional Exit button to Pads toggle menu.  (#1)
* Add IsValidEntity check to Sound Hook

In some cases, this will attempt to run on invalid entities, resulting in error log spam.

* Add Cvar "pads_menu_exit_type" for Menu Exit Button

I don't know if it was intentional for having the exit button disabled. But, some hosts (or clients), might be bugged that you cannot close this menu freely without having to pick an option. So it's now an optional to keep on or off.